requirements.txt
pdfplumber
mammoth
markdownify
python-docx
flake8
black
pre-commit
pytest
